pyproject.toml
[project]
name = "memoriesdb"
version = "0.2.0"
description = "MemoriesDB, a graph/vector/neural database"
readme = "README.md"
authors = [
    { name = "@val315159", email = "jmward@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=0.19.0",
    "honcho>=2.0.0",
    "ollama>=0.5.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "gevent-websocket>=0.10.1",
    "websocket-client>=1.8.0",
    "bottle-dev>=0.13.2",
    "uvicorn>=0.35.0",
    "httpx>=0.28.1",
    "fastapi>=0.115.14",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.1",
    "uv>=0.7.12",
    "Faker>=13.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
